openssd: add append point structure
For each pool available within a device, we define 1..X append
points. An append point is a place to write within a block of a
device.

If there is more than 1 append point in a pool. They have to wait
for writing to the pool. i.e. we assume that a pool serialize all
accesses.

This let us model channels within a flash SSD, of which we can
only have one write/read/erase running within a chip.

The model is kind of simplified to fit the our model. A better
simulation can be added if needed at a later point.