openssd: add logical locking data structures
We need to lock a logical flash block when we wish to perform some
form of GC in the backgroup. For this we apply a state to each
block, that allows us to know the state of each block.

The state is only changed when an block is starting to be repurposed,
for future writes, thus its administration is removed from the hot
path.