vsl: Implemented debug-only sanity check abilities.
BUG_ON is always evaluated, regardless of compiling the kernel for
production or debugging/development use - hence the modified make
file and vsl.h definition of 'VSL_ASSERT(c)', used for validating
preconditions in code provided DEBUG=y is set in the makefile.

Example: can be used to ensure functions requiring locks to be
taken aren't invoked unless this is the case:

--
VSL_ASSERT(spin_is_locked(&block->lock));
--

The difference lies in the lack of overhead on production code.