vsl: Let device driver issue requests
Previously, vsl executed the request on behalf of the device driver. To
integrate easily into the scsi stack, we instead let vsl map the request
and return the changed req for the device driver to execute it.

We should put in some way to detect that there's a mapping associated
with it.