vsl: progress on key-value store
The code is very rudimentary and subject to imminent, large-scale
changes - but it hints at the final form.
One large, fixed size hash with 0.25 bucket occupancy and 16
entries per bucket - h1(key) modulo the number of buckets
determines the resulting bucket, and each entry will contain a 16B
MD5 hash of the key (h2(key)) to further reduce the likelyhood of
genuine collisions while still not retaining the full key.